+++ integrationTest/test/jasmine-jasmine/verify/verify.ts
@@ -1,15 +1,10 @@
 import * as fs from 'mz/fs';
 import { expect } from 'chai';
-import * as path from 'path';
-import { ScoreResult } from 'stryker-api/report';
+import { readScoreResult } from '../../../helpers';
 
 describe('After running stryker with test runner jasmine, test framework jasmine', () => {
   it('should report 85% mutation score', async () => {
-    const allReportFiles = await fs.readdir(path.resolve('reports', 'mutation', 'events'));
-    const scoreResultReportFile = allReportFiles.find(file => !!file.match(/.*onScoreCalculated.*/));
-    expect(scoreResultReportFile).ok;
-    const scoreResultContent = await fs.readFile(path.resolve('reports', 'mutation', 'events', scoreResultReportFile || ''), 'utf8');
-    const scoreResult = JSON.parse(scoreResultContent) as ScoreResult;
+    const scoreResult = await readScoreResult();
     expect(scoreResult.killed).eq(12);
     expect(scoreResult.noCoverage).eq(1);
     expect(scoreResult.mutationScore).greaterThan(85).and.lessThan(86);
@@ -18,7 +13,7 @@ describe('After running stryker with test runner jasmine, test framework jasmine
   it('should write to a log file', async () => {
     const strykerLog = await fs.readFile('./stryker.log', 'utf8');
     expect(strykerLog).contains('INFO InputFileResolver Found 2 of 10 file(s) to be mutated');
-    expect(strykerLog).matches(/Stryker Done in \d+ seconds/);
+    expect(strykerLog).matches(/Stryker Done in \d+/);
     // TODO, we now have an error because of a memory leak: https://github.com/jasmine/jasmine-npm/issues/134
     // expect(strykerLog).not.contains('ERROR'); 
   });