+++ packages/stryker-vue-mutator/src/index.ts
@@ -1,0 +1,4 @@
+import { MutatorFactory } from 'stryker-api/mutant';
+import VueMutator from './VueMutator';
+
+MutatorFactory.instance().register('vue', VueMutator);