+++ packages/stryker-vue-mutator/test/helpers/globals.ts
@@ -1,0 +1,6 @@
+declare const sandbox: sinon.SinonSandbox;
+namespace NodeJS {
+  export interface Global {
+    sandbox: sinon.SinonSandbox;
+  }
+}