+++ packages/stryker/src/test-runner/TimeoutDecorator.ts
@@ -1,8 +1,7 @@
 import { RunOptions, RunResult, RunStatus } from 'stryker-api/test_runner';
 import TestRunnerDecorator from './TestRunnerDecorator';
 import { getLogger } from 'stryker-api/logging';
-import { TIMEOUT_EXPIRED, timeout } from '../utils/objectUtils';
-
+import { TimeoutExpired, timeout } from '../utils/objectUtils';
 
 /**
  * Wraps a test runner and implements the timeout functionality.
@@ -14,7 +13,7 @@ export default class TimeoutDecorator extends TestRunnerDecorator {
   async run(options: RunOptions): Promise<RunResult> {
     this.log.debug('Starting timeout timer (%s ms) for a test run', options.timeout);
     const result = await timeout(super.run(options), options.timeout);
-    if (result === TIMEOUT_EXPIRED) {
+    if (result === TimeoutExpired) {
       return this.handleTimeout();
     } else {
       return result;