+++ packages/stryker/testResources/command-runner/ok.js
@@ -1,0 +1,1 @@
+console.log('Test 1 - OK')