+++ packages/stryker-karma-runner/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.16.5"></a>
+## [0.16.5](https://github.io/stryker-mutator/stryker/compare/stryker-karma-runner@0.16.4...stryker-karma-runner@0.16.5) (2018-08-17)
+
+
+### Bug Fixes
+
+* **dependencies:** support stryker-api 0.19.0 ([#1087](https://github.io/stryker-mutator/stryker/issues/1087)) ([44ce923](https://github.io/stryker-mutator/stryker/commit/44ce923))
+
+
+
+
+<a name="0.16.4"></a>
+## [0.16.4](https://github.io/stryker-mutator/stryker/compare/stryker-karma-runner@0.16.3...stryker-karma-runner@0.16.4) (2018-08-17)
+
+
+
+
+**Note:** Version bump only for package stryker-karma-runner
+
+<a name="0.16.3"></a>
+## [0.16.3](https://github.io/stryker-mutator/stryker/compare/stryker-karma-runner@0.16.2...stryker-karma-runner@0.16.3) (2018-08-16)
+
+
+
+
+**Note:** Version bump only for package stryker-karma-runner
+
 <a name="0.16.2"></a>
 ## [0.16.2](https://github.io/stryker-mutator/stryker/compare/stryker-karma-runner@0.16.1...stryker-karma-runner@0.16.2) (2018-08-03)