+++ packages/stryker-mocha-framework/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.11.2"></a>
+## [0.11.2](https://github.com/stryker-mutator/stryker/compare/stryker-mocha-framework@0.11.1...stryker-mocha-framework@0.11.2) (2018-08-17)
+
+
+### Bug Fixes
+
+* **dependencies:** support stryker-api 0.19.0 ([#1087](https://github.com/stryker-mutator/stryker/issues/1087)) ([44ce923](https://github.com/stryker-mutator/stryker/commit/44ce923))
+
+
+
+
+<a name="0.11.1"></a>
+## [0.11.1](https://github.com/stryker-mutator/stryker/compare/stryker-mocha-framework@0.11.0...stryker-mocha-framework@0.11.1) (2018-08-17)
+
+
+
+
+**Note:** Version bump only for package stryker-mocha-framework
+
 <a name="0.11.0"></a>
 # [0.11.0](https://github.com/stryker-mutator/stryker/compare/stryker-mocha-framework@0.10.3...stryker-mocha-framework@0.11.0) (2018-07-20)